Mean Variance Optimizer in Python
Warning: The package pyfolio works in Google Colab, but not in WinPython.

Because:
- Create a simple 1/n strategy portfolio tearsheet in Python

The portfolio is based on a Hackernoon article:

URL: https://towardsdatascience.com/algorithmic-trading-based-on-mean-variance-optimization-in-python-62bdf844ac5b

This commit does not resolve the portfolio optimizer, as it only
creates a basic tearsheet using pyfolio for the 1/n strategy. Other
portfolio strategies are not implemented as the article skipped over
large portions of the code.

The article is extracted from the book Python for Finance by Eryk Lewinson.