PIECE analysis in Python
Because:
- PIECE stands for Profit, Inflow of Operational Cash, Return on Equity,
Conservative Debt, and Economic Moat
- Fetches fundamental data from StockRow
- Fetches description and technical data from IEX Cloud
- Plots EMA wave chart
- Allows multiple tickers

Third-party dependencies:
- StockRow Excel spreadsheets
- IEX Cloud (requires token)

This code is complete for a manual PIECE analysis, as it fetches the
US stock description, fundamental data and price from sources. User
needs to analyze the fundamental data that is fetched, because no
guidance is given.

In practice, we give a plus rating for each:
- Increasing Annual Revenue > $750M
- Increasing Operational Cashflow > $100M
- Increasing ROE > 15%
- Decreasing Debt-to-Equity < 0.5

Economic Moat cannot be measured empirically, but user has to analyze
stocks further, e.g. SeekingAlpha.

This commit does not include automated PIECE analysis and data is not
pretty to look at.

Because:
- Does not use ML to automatically rate companies based on historical
revenue, operating cash flow, ROE, and Debt-to-Equity
- Does not have Singapore stocks, try alternative source UncleStock
- Presentation of data is raw and not pretty to look at.